Reformat + cleanup. Move basic logger to testing dir. Cleanup actuator files. Cleanup analyser files. Cleanup sensor files. Cleanup main.py.
