+++ src/analyser/brightness_analyser.py
@@ -25,5 +25,3 @@ class BrightnessAnalyser(Analyser):
         else:
             sensor_data.actuator_value = 0
             pub.sendMessage("actuator.light_status", args=sensor_data)
-            # FailsafeDefault value (for safety reasons,
-            # keep it as 0, as per our security module :)
