+++ src/data_streamer/.env.template
